Dependency of Keyboard on wxKeyEvent moved to MenuCreator
