+++ src/MenuCreator.h
@@ -66,4 +66,9 @@ private:
    Observer::Subscription mUndoSubscription;
 };
 
+struct NormalizedKeyString;
+
+AUDACITY_DLL_API
+NormalizedKeyString KeyEventToKeyString(const wxKeyEvent & keyEvent);
+
 #endif
