+++ src/commands/Keyboard.h
@@ -34,7 +34,4 @@ namespace std
       : hash< NormalizedKeyStringBase > {};
 }
 
-AUDACITY_DLL_API
-NormalizedKeyString KeyEventToKeyString(const wxKeyEvent & keyEvent);
-
 #endif
